Put attribute preparation in separate class
