+++ src/main/java/eu/dissco/core/handlemanager/service/FdoRecordService.java
@@ -1,0 +1,260 @@
+package eu.dissco.core.handlemanager.service;
+
+import static eu.dissco.core.handlemanager.domain.PidRecords.DIGITAL_OBJECT_SUBTYPE;
+import static eu.dissco.core.handlemanager.domain.PidRecords.DIGITAL_OBJECT_TYPE;
+import static eu.dissco.core.handlemanager.domain.PidRecords.DIGITAL_OR_PHYSICAL;
+import static eu.dissco.core.handlemanager.domain.PidRecords.FIELD_IDX;
+import static eu.dissco.core.handlemanager.domain.PidRecords.HS_ADMIN;
+import static eu.dissco.core.handlemanager.domain.PidRecords.IN_COLLECTION_FACILITY;
+import static eu.dissco.core.handlemanager.domain.PidRecords.ISSUE_DATE;
+import static eu.dissco.core.handlemanager.domain.PidRecords.ISSUE_NUMBER;
+import static eu.dissco.core.handlemanager.domain.PidRecords.LOC;
+import static eu.dissco.core.handlemanager.domain.PidRecords.LOC_REQ;
+import static eu.dissco.core.handlemanager.domain.PidRecords.MEDIA_HASH;
+import static eu.dissco.core.handlemanager.domain.PidRecords.MEDIA_URL;
+import static eu.dissco.core.handlemanager.domain.PidRecords.NODE_ID;
+import static eu.dissco.core.handlemanager.domain.PidRecords.OBJECT_TYPE;
+import static eu.dissco.core.handlemanager.domain.PidRecords.PHYSICAL_IDENTIFIER;
+import static eu.dissco.core.handlemanager.domain.PidRecords.PID;
+import static eu.dissco.core.handlemanager.domain.PidRecords.PID_ISSUER;
+import static eu.dissco.core.handlemanager.domain.PidRecords.PID_KERNEL_METADATA_LICENSE;
+import static eu.dissco.core.handlemanager.domain.PidRecords.PID_STATUS;
+import static eu.dissco.core.handlemanager.domain.PidRecords.PRESERVED_OR_LIVING;
+import static eu.dissco.core.handlemanager.domain.PidRecords.REFERENT;
+import static eu.dissco.core.handlemanager.domain.PidRecords.REFERENT_DOI_NAME;
+import static eu.dissco.core.handlemanager.domain.PidRecords.SPECIMEN_HOST;
+import static eu.dissco.core.handlemanager.domain.PidRecords.SUBJECT_PHYSICAL_IDENTIFIER;
+import static eu.dissco.core.handlemanager.domain.PidRecords.SUBJECT_SPECIMEN_HOST;
+import static eu.dissco.core.handlemanager.service.ServiceUtils.setUniquePhysicalIdentifierId;
+import static eu.dissco.core.handlemanager.utils.AdminHandleGenerator.genAdminHandle;
+
+import eu.dissco.core.handlemanager.domain.repsitoryobjects.HandleAttribute;
+import eu.dissco.core.handlemanager.domain.requests.attributes.DigitalSpecimenBotanyRequest;
+import eu.dissco.core.handlemanager.domain.requests.attributes.DigitalSpecimenRequest;
+import eu.dissco.core.handlemanager.domain.requests.attributes.DoiRecordRequest;
+import eu.dissco.core.handlemanager.domain.requests.attributes.HandleRecordRequest;
+import eu.dissco.core.handlemanager.domain.requests.attributes.MediaObjectRequest;
+import eu.dissco.core.handlemanager.exceptions.PidServiceInternalError;
+import java.io.StringWriter;
+import java.nio.charset.StandardCharsets;
+import java.time.Instant;
+import java.time.ZoneId;
+import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Locale;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.transform.OutputKeys;
+import javax.xml.transform.TransformerException;
+import javax.xml.transform.TransformerFactory;
+import javax.xml.transform.dom.DOMSource;
+import javax.xml.transform.stream.StreamResult;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.w3c.dom.Document;
+
+@Service
+@RequiredArgsConstructor
+public class FdoRecordService {
+
+  private final TransformerFactory tf;
+  private final DocumentBuilderFactory dbf;
+
+  private final DateTimeFormatter dt = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS",
+      Locale.ENGLISH).withZone(ZoneId.of("UTC"));
+
+  protected List<HandleAttribute> prepareHandleRecordAttributes(HandleRecordRequest request,
+      byte[] handle) throws PidServiceInternalError {
+    List<HandleAttribute> fdoRecord = new ArrayList<>();
+
+    // 100: Admin Handle
+    fdoRecord.add(
+        new HandleAttribute(FIELD_IDX.get(HS_ADMIN), handle, HS_ADMIN, genAdminHandle()));
+
+    // 1: Pid
+    byte[] pid = ("https://hdl.handle.net/" + new String(handle, StandardCharsets.UTF_8)).getBytes(
+        StandardCharsets.UTF_8);
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(PID), handle, PID, pid));
+
+    // 2: PidIssuer
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(PID_ISSUER), handle, PID_ISSUER,
+        request.getPidIssuerPid().getBytes(StandardCharsets.UTF_8)));
+
+    // 3: Digital Object Type
+    fdoRecord.add(
+        new HandleAttribute(FIELD_IDX.get(DIGITAL_OBJECT_TYPE), handle, DIGITAL_OBJECT_TYPE,
+            request.getDigitalObjectTypePid().getBytes(StandardCharsets.UTF_8)));
+
+    // 4: Digital Object Subtype
+    fdoRecord.add(
+        new HandleAttribute(FIELD_IDX.get(DIGITAL_OBJECT_SUBTYPE), handle, DIGITAL_OBJECT_SUBTYPE,
+            request.getDigitalObjectSubtypePid().getBytes(StandardCharsets.UTF_8)));
+
+    // 5: 10320/loc
+    byte[] loc = setLocations(request.getLocations(), new String(handle, StandardCharsets.UTF_8));
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(LOC), handle, LOC, loc));
+
+    // 6: Issue Date
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(ISSUE_DATE), handle, ISSUE_DATE,
+        getDate().getBytes(StandardCharsets.UTF_8)));
+
+    // 7: Issue number
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(ISSUE_NUMBER), handle, ISSUE_NUMBER,
+        "1".getBytes(StandardCharsets.UTF_8)));
+
+    // 8: PidStatus
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(PID_STATUS), handle, PID_STATUS,
+        "TEST".getBytes(StandardCharsets.UTF_8)));
+
+    // 9, 10: tombstone text, tombstone pids -> Skip
+
+    // 11: PidKernelMetadataLicense:
+    byte[] pidKernelMetadataLicense = "https://creativecommons.org/publicdomain/zero/1.0/".getBytes(
+        StandardCharsets.UTF_8);
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(PID_KERNEL_METADATA_LICENSE), handle,
+        PID_KERNEL_METADATA_LICENSE, pidKernelMetadataLicense));
+
+    return fdoRecord;
+  }
+
+  protected List<HandleAttribute> prepareDoiRecordAttributes(DoiRecordRequest request, byte[] handle)
+      throws PidServiceInternalError {
+    var fdoRecord = prepareHandleRecordAttributes(request, handle);
+
+    // 12: Referent DOI Name
+    fdoRecord.add(
+        new HandleAttribute(FIELD_IDX.get(REFERENT_DOI_NAME), handle, REFERENT_DOI_NAME,
+            request.getReferentDoiNamePid().getBytes(StandardCharsets.UTF_8)));
+
+    // 13: Referent -> NOTE: Referent is blank currently until we have a model
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(REFERENT), handle, REFERENT,
+        request.getReferent().getBytes(StandardCharsets.UTF_8)));
+    return fdoRecord;
+  }
+
+  protected List<HandleAttribute> prepareMediaObjectAttributes(MediaObjectRequest request,
+      byte[] handle)
+      throws PidServiceInternalError {
+    var fdoRecord = prepareDoiRecordAttributes(request, handle);
+
+    // 14 Media Hash
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(MEDIA_HASH), handle, MEDIA_HASH,
+        request.getMediaHash().getBytes(StandardCharsets.UTF_8)));
+
+    // 15 Subject Specimen Host
+    fdoRecord.add(
+        new HandleAttribute(FIELD_IDX.get(SUBJECT_SPECIMEN_HOST), handle, SUBJECT_SPECIMEN_HOST,
+            request.getSubjectSpecimenHostPid().getBytes(StandardCharsets.UTF_8)));
+
+    // 16 Media Url
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(MEDIA_URL), handle, MEDIA_URL,
+        request.getMediaUrl().getBytes(StandardCharsets.UTF_8)));
+
+    // 17 : Subject Physical Identifier
+    // Encoding here is UTF-8
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(SUBJECT_PHYSICAL_IDENTIFIER), handle,
+        SUBJECT_PHYSICAL_IDENTIFIER, setUniquePhysicalIdentifierId(request)));
+    return fdoRecord;
+  }
+
+  protected List<HandleAttribute> prepareDigitalSpecimenRecordAttributes(
+      DigitalSpecimenRequest request, byte[] handle)
+      throws PidServiceInternalError {
+    var fdoRecord = prepareDoiRecordAttributes(request, handle);
+
+    fdoRecord.add(
+        new HandleAttribute(FIELD_IDX.get(DIGITAL_OR_PHYSICAL), handle, DIGITAL_OR_PHYSICAL,
+            request.getDigitalOrPhysical().getBytes()));
+
+    // 15: specimenHost
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(SPECIMEN_HOST), handle, SPECIMEN_HOST,
+        request.getSpecimenHostPid().getBytes(StandardCharsets.UTF_8)));
+
+    // 16: In collectionFacility
+    fdoRecord.add(
+        new HandleAttribute(FIELD_IDX.get(IN_COLLECTION_FACILITY), handle, IN_COLLECTION_FACILITY,
+            request.getInCollectionFacilityPid().getBytes(StandardCharsets.UTF_8)));
+
+    // 17 : Institutional Identifier
+    // Encoding here is UTF-8
+    fdoRecord.add(
+        new HandleAttribute(FIELD_IDX.get(PHYSICAL_IDENTIFIER), handle, PHYSICAL_IDENTIFIER,
+            setUniquePhysicalIdentifierId(request)));
+
+    return fdoRecord;
+  }
+
+  protected List<HandleAttribute> prepareDigitalSpecimenBotanyRecordAttributes(
+      DigitalSpecimenBotanyRequest request, byte[] handle)
+      throws PidServiceInternalError {
+    List<HandleAttribute> fdoRecord = prepareDigitalSpecimenRecordAttributes(request, handle);
+
+    // 17: ObjectType
+    fdoRecord.add(new HandleAttribute(FIELD_IDX.get(OBJECT_TYPE), handle, OBJECT_TYPE,
+        request.getObjectType().getBytes(StandardCharsets.UTF_8)));
+
+    // 18: preservedOrLiving
+    fdoRecord.add(
+        new HandleAttribute(FIELD_IDX.get(PRESERVED_OR_LIVING), handle, PRESERVED_OR_LIVING,
+            request.getPreservedOrLiving().getBytes()));
+
+    return fdoRecord;
+  }
+
+  private String getDate() {
+    return dt.format(Instant.now());
+  }
+
+  public byte[] setLocations(String[] userLocations, String handle) throws PidServiceInternalError {
+
+    DocumentBuilder documentBuilder;
+    try {
+      documentBuilder = dbf.newDocumentBuilder();
+    } catch (ParserConfigurationException e) {
+      throw new PidServiceInternalError(e.getMessage(), e);
+    }
+
+    var doc = documentBuilder.newDocument();
+    var locations = doc.createElement(LOC_REQ);
+    doc.appendChild(locations);
+    String[] objectLocations = concatLocations(userLocations, handle);
+
+    for (int i = 0; i < objectLocations.length; i++) {
+      var locs = doc.createElement("location");
+      locs.setAttribute(NODE_ID, String.valueOf(i));
+      locs.setAttribute("href", objectLocations[i]);
+      String weight = i < 1 ? "1" : "0";
+      locs.setAttribute("weight", weight);
+      locations.appendChild(locs);
+    }
+    try {
+      return documentToString(doc).getBytes(StandardCharsets.UTF_8);
+    } catch (TransformerException e) {
+      throw new PidServiceInternalError("An internal error has occurred parsing location data", e);
+    }
+  }
+
+  private String[] concatLocations(String[] userLocations, String handle){
+    ArrayList<String> objectLocations = new ArrayList<>(List.of(defaultLocations(handle)));
+    if (userLocations != null){
+      objectLocations.addAll(List.of(userLocations));
+    }
+    return objectLocations.toArray(new String[0]);
+  }
+
+  private String[] defaultLocations(String handle){
+    String api = "https://sandbox.dissco.tech/api/v1/specimens/" + handle;
+    String ui = "https://sandbox.dissco.tech/ds/" + handle;
+    return new String[]{api, ui};
+  }
+
+  private String documentToString(Document document) throws TransformerException {
+    var transformer = tf.newTransformer();
+    transformer.setOutputProperty(OutputKeys.OMIT_XML_DECLARATION, "yes");
+    StringWriter writer = new StringWriter();
+    transformer.transform(new DOMSource(document), new StreamResult(writer));
+    return writer.getBuffer().toString();
+  }
+}
